+++ src/lib/abstract-collection-form-field/abstract-collection-form-field.component.ts
@@ -1,24 +1,32 @@
-import {FormArray, FormControl} from '@angular/forms';
 import {AbstractCollectionType} from '../form-type/abstract-collection-type';
 import {AbstractFormFieldComponent} from '../abstract-form-field/abstract-form-field.component';
-import {AbstractGroupType} from '../form-type/abstract-group-type';
 import {ModelHandler} from '../model-handler';
+import {FormArray} from '../abstraction';
+import {Input} from '@angular/core';
 
+// noinspection JSUnusedGlobalSymbols
 export abstract class AbstractCollectionFormFieldComponent extends AbstractFormFieldComponent<AbstractCollectionType<any, any>> {
 
+  @Input()
+  public element: FormArray;
+
+  // noinspection JSUnusedGlobalSymbols
   public removeEntry(index: number): void {
-    (this.element as FormArray).removeAt(index);
+    this.element.removeAt(index);
   }
 
+  // noinspection JSUnusedGlobalSymbols
   public addEntry(): void {
-    if (this.fieldType.fieldInstance instanceof AbstractGroupType) {
-      const field = ModelHandler.build(this.fieldType.fieldInstance.options.model, this.fieldType.builderInstance);
-      field.setParent(this.element as FormArray);
-      (this.element as FormArray).controls.push(field);
-    } else if (this.fieldType.fieldInstance instanceof AbstractCollectionType) {
-      (this.element as FormArray).controls.push(new FormArray([], this.fieldType.fieldInstance.validators));
-    } else {
-      (this.element as FormArray).controls.push(new FormControl(null, this.fieldType.fieldInstance.validators));
+    const newControl = ModelHandler.buildSingleField(
+      this.fieldType.fieldInstance,
+      this.fieldType.builderInstance
+    );
+
+    if (!newControl) {
+      return;
     }
+
+    newControl.initHandler.setIsInitialized(true);
+    this.element.controls.push(newControl);
   }
 }