+++ src/lib/form-builder/form-builder.component.ts
@@ -1,10 +1,22 @@
-import {Component, ContentChildren, EventEmitter, Input, OnChanges, OnInit, Output, QueryList, SimpleChanges} from '@angular/core';
-import {FormArray, FormControl, FormGroup, ValidationErrors} from '@angular/forms';
+import {
+  Component,
+  ContentChildren,
+  EventEmitter,
+  Input,
+  OnChanges,
+  OnInit,
+  Output,
+  QueryList,
+  SimpleChanges
+} from '@angular/core';
+import {ValidationErrors} from '@angular/forms';
 import {AbstractType} from '../form-type/abstract-type';
 import {FormSlotComponent} from '../form-slot/form-slot.component';
 import {ModelHandler} from '../model-handler';
 import {AbstractGroupType} from '../form-type/abstract-group-type';
 import {AbstractCollectionType} from '../form-type/abstract-collection-type';
+import {FormArray, FormControl, FormGroup} from '../abstraction';
+import {AbstractFormControl} from '../types';
 
 export interface FormModel {
   [key: string]: AbstractType<any>;
@@ -14,23 +26,27 @@ export interface FormModel {
   selector: 'mw-form-builder',
   template: `
     <ng-content></ng-content>
-    <form [formGroup]="group" (ngSubmit)="submit()" #form>
-      <mw-form-group #fromGroup [element]="group" [formGroup]="group" [model]="formModel" [slots]="slots">
+    <form [formGroup]="group" (ngSubmit)="submit()">
+      <mw-form-group [element]="group"
+                     [formGroup]="group"
+                     [model]="formModel"
+                     [slots]="slots"
+                     [isRootGroup]="true">
       </mw-form-group>
     </form>
   `,
   styles: []
 })
-export class FormBuilderComponent implements OnInit, OnChanges {
+export class FormBuilderComponent<T extends { [key: string]: any } = {}> implements OnInit, OnChanges {
 
   @Input()
   public formModel: FormModel;
 
   @Input()
-  public formData: { [key: string]: any };
+  public formData: T;
 
   @Output()
-  public onSubmit: EventEmitter<any> = new EventEmitter<any>();
+  public onSubmit: EventEmitter<T> = new EventEmitter<T>();
 
   @ContentChildren(FormSlotComponent, {descendants: true})
   public slots: QueryList<FormSlotComponent>;
@@ -38,35 +54,44 @@ export class FormBuilderComponent implements OnInit, OnChanges {
   public group: FormGroup;
 
   public ngOnInit(): void {
-    this.rebuildForm();
+    if (this.group == null) {
+      this.rebuildForm();
+    }
   }
 
   public ngOnChanges(changes: SimpleChanges): void {
-    if (changes.hasOwnProperty('formModel') && changes.formModel.currentValue != null) {
+    if ('formModel' in changes && changes.formModel.currentValue) {
       this.rebuildForm();
     }
-    if (changes.hasOwnProperty('formData') && changes.formData.currentValue != null) {
-      if (!changes.hasOwnProperty('formModel')) {
+    if ('formData' in changes && changes.formData.currentValue) {
+      if (!('formModel' in changes)) {
         this.initializeCollectionFields(this.group, this.formModel, this.formData);
+        this.group.initHandler.setIsInitialized(false);
+        this.group.initHandler.setIsInitialized(true);
       }
-      if (this.group != null) {
+      if (this.group) {
         this.group.patchValue(this.formData);
       }
     }
   }
 
   public rebuildForm(): void {
-    let res = {};
-    if (this.group != null) {
+    let res = null;
+    let groupIsInitialized = false;
+    if (this.group) {
       res = Object.assign({}, this.group.value);
-    } else if (this.formData != null) {
+      groupIsInitialized = this.group.initHandler.isInitialized;
+    } else if (this.formData) {
       res = Object.assign({}, this.formData);
     }
 
-    this.buildForm();
+    this.group = ModelHandler.build(this.formModel, this);
     this.initializeCollectionFields(this.group, this.formModel, res);
 
-    if (this.group != null) {
+    if (this.group && res) {
+      if (groupIsInitialized) {
+        this.group.initHandler.setIsInitialized(true);
+      }
       this.group.patchValue(res);
     }
   }
@@ -74,66 +99,54 @@ export class FormBuilderComponent implements OnInit, OnChanges {
   private initializeCollectionFields(group: FormGroup, model: FormModel, data: any): void {
     Object.keys(model).forEach((name) => {
       const formType = model[name];
-      const childData = (data != null && data.hasOwnProperty(name)) ? data[name] : null;
+      const childData = (data && name in data) ? data[name] : null;
 
       const control = group.get(name);
-      if (formType instanceof AbstractGroupType) {
-        this.initializeCollectionFields(control as FormGroup, (formType as AbstractGroupType<any>).options.model, childData);
-      } else if (formType instanceof AbstractCollectionType) {
-        this.initializeCollectionField(formType, control as FormArray, childData);
+      if (formType instanceof AbstractGroupType && control instanceof FormGroup) {
+        this.initializeCollectionFields(control, formType.options.model, childData);
+      } else if (formType instanceof AbstractCollectionType && control instanceof FormArray) {
+        this.initializeCollectionField(formType, control, childData);
       }
     });
   }
 
   private initializeCollectionField(collection: AbstractCollectionType<any, any>, array: FormArray, data?: Array<any>): void {
     const numberOfEntries = Array.isArray(data) ? data.length : 0;
-    [...new Array(array.controls.length).keys()].forEach((n) => array.removeAt(n));
+
+    array.controls.splice(0, array.controls.length);
     if (numberOfEntries === 0) {
       return;
     }
 
     const fieldInstance = collection.fieldInstance;
-    const isGroup = fieldInstance instanceof AbstractGroupType;
-    const isCollection = !isGroup && fieldInstance instanceof AbstractCollectionType;
-    const isFormField = !isCollection && fieldInstance instanceof AbstractType;
 
     for (let i = 0; i < numberOfEntries; i++) {
-      if (isGroup) {
-        const childModel = (fieldInstance as AbstractGroupType<any>).options.model;
-        const childFormGroup = ModelHandler.build(childModel, this);
-        childFormGroup.setParent(array);
-
-        if (fieldInstance.disabled) {
-          childFormGroup.disable();
-        }
-
-        this.initializeCollectionFields(childFormGroup, childModel, data[i]);
-        array.controls.push(childFormGroup);
-      } else if (isCollection) {
-        const childArray = new FormArray([], fieldInstance.validators);
-        childArray.setParent(array);
-
-        if (fieldInstance.disabled) {
-          childArray.disable();
-        }
-
-        this.initializeCollectionField(fieldInstance as AbstractCollectionType<any, any>, childArray, data[i]);
-        array.controls.push(childArray);
-      } else if (isFormField) {
-        array.controls.push(new FormControl({value: null, disabled: fieldInstance.disabled}, fieldInstance.validators));
+      const control: AbstractFormControl = ModelHandler.buildSingleField(fieldInstance, this);
+      if (!control) {
+        continue;
       }
+      if (fieldInstance instanceof AbstractGroupType && control instanceof FormGroup) {
+        control.setParent(array);
+
+        if (fieldInstance.disabled) {
+          control.disable();
+        }
+
+        this.initializeCollectionFields(control, fieldInstance.options.model, data[i]);
+      } else if (fieldInstance instanceof AbstractCollectionType && control instanceof FormArray) {
+        control.setParent(array);
+
+        if (fieldInstance.disabled) {
+          control.disable();
+        }
+
+        this.initializeCollectionField(fieldInstance, control, data[i]);
+      }
+      array.controls.push(control);
     }
   }
 
-  /**
-   * This method builds the FormGroup
-   */
-  private buildForm(): void {
-    this.group = ModelHandler.build(this.formModel, this);
-  }
-
-
-  public submit(): any {
+  public submit(): T {
     Object.keys(this.group.controls).forEach((field) => {
       const control = this.group.get(field);
       control.markAsDirty();
@@ -161,8 +174,8 @@ export class FormBuilderComponent implements OnInit, OnChanges {
       idx++;
       const control = group.get(name);
       if (control instanceof FormArray || control instanceof FormGroup) {
-        const tmpErrors = this.getErrors(control as (FormArray | FormGroup));
-        if (tmpErrors != null) {
+        const tmpErrors = this.getErrors(control);
+        if (tmpErrors) {
           errors[name] = tmpErrors;
         }
       } else if (control instanceof FormControl) {