+++ src/lib/model-handler.ts
@@ -1,8 +1,8 @@
 import {FormBuilderComponent, FormModel} from './form-builder/form-builder.component';
-import {AbstractControl, FormArray, FormControl, FormGroup} from '@angular/forms';
 import {AbstractCollectionType} from './form-type/abstract-collection-type';
 import {AbstractGroupType} from './form-type/abstract-group-type';
 import {AbstractType} from './form-type/abstract-type';
+import {FormArray, FormControl, FormGroup} from './abstraction';
 
 export class ModelHandler {
   public static build(model: FormModel, builderInstance: FormBuilderComponent): FormGroup {
@@ -14,8 +14,11 @@ export class ModelHandler {
   private static handleModel(model: FormModel, group: FormGroup, builderInstance: FormBuilderComponent): void {
     Object.keys(model).forEach((name) => {
       const field = model[name];
-      if (field != null) {
-        const control = ModelHandler.handleField(field, builderInstance);
+      if (field) {
+        const control = ModelHandler.buildSingleField(field, builderInstance);
+        if (!control) {
+          return;
+        }
         control.setParent(group);
 
         if (field.disabled) {
@@ -23,28 +26,37 @@ export class ModelHandler {
         }
 
         group.addControl(name, control);
-        if (field.builderInstance == null && builderInstance != null) {
+        if (field.builderInstance == null && builderInstance) {
           field.builderInstance = builderInstance;
         }
+
+        field.control = control;
       }
     });
   }
 
-  private static handleField(field: AbstractType<any>, builderInstance: FormBuilderComponent): AbstractControl {
-    let component: AbstractControl;
+  public static buildSingleField(field: AbstractType<any>,
+                                 builderInstance: FormBuilderComponent
+  ): FormArray | FormGroup | FormControl | null {
+    let component: FormArray | FormGroup | FormControl;
+
     if (field instanceof AbstractCollectionType) {
-      component = new FormArray([], field.validators);
-      ModelHandler.handleArray(field as AbstractCollectionType<any, any>, component as FormArray, builderInstance);
+      component = new FormArray([], field.controlOptions);
+      ModelHandler.handleArray(field, component, builderInstance);
     } else if (field instanceof AbstractGroupType) {
-      component = new FormGroup({}, field.validators);
-      ModelHandler.handleModel(field.options.model, component as FormGroup, builderInstance);
-    } else {
-      component = new FormControl({value: null, disabled: field.disabled}, field.validators);
+      component = new FormGroup({}, field.controlOptions);
+      ModelHandler.handleModel(field.options.model, component, builderInstance);
+    } else if (field instanceof AbstractType) {
+      component = new FormControl({value: null, disabled: field.disabled}, field.controlOptions);
     }
     return component;
   }
 
   private static handleArray(field: AbstractCollectionType<any, any>, array: FormArray, builderInstance: FormBuilderComponent): void {
-    array.push(ModelHandler.handleField(field.fieldInstance, builderInstance));
+    const control = ModelHandler.buildSingleField(field.fieldInstance, builderInstance);
+    if (!control) {
+      return;
+    }
+    array.push(control);
   }
 }